pyproject.toml
[project]
name = "credit-contract-backend"
version = "1.0.0"
description = "Backend API for personal credit contract management."
authors = [
    { name = "pleitemarques", email = "p22dro@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "django>=5.2.2,<6.0.0",
    "djangorestframework>=3.16.0,<4.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "django-filter>=25.1,<26.0",
    "drf-spectacular>=0.28.0,<0.29.0",
    "python-decouple>=3.8,<4.0",
    "django-cors-headers>=4.7.0,<5.0.0",
    "django-oauth-toolkit (>=3.0.1,<4.0.0)",
    "validate-docbr (>=1.11.0,<2.0.0)",
    "phonenumbers (>=9.0.7,<10.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
coverage = "^7.8.2"
ruff = "^0.11.13"
black = "^25.1.0"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"